requirements.txt
numpy>=1.24.0,<2.0.0
faiss-cpu>=1.7.4
umap-learn==0.5.7
matplotlib==3.10.1
seaborn==0.13.2
sentence-transformers==4.0.2
emoji==2.14.1
h5py==3.13.0
openai==1.72.0
python-dotenv==1.1.0